Create tracking.php
This simple script was written quite a few years ago, but it is here to demonstrate I have used PHP and mySQL.

I can't remember all the details now, but the purpose of this script was to log information about the users who visited my websites to a MySQL database. Information I wanted to log was IP address, referrer, user agent, the domain (which of my sites did they visit) and the page visited. The page visited corresponded to a single keyword so it helped me to see what subject matter was attracting the most vistors.